simplify randomization of photos and answers
NOTE: i haven't been able to compile this on linux due to some missing
openGL and x11 libs so it may not compile at all.

i just wanted to show how we might be able to simplify the randomization
a little.  i developed this with some tests and seemed to be producing
good looking numbers.  but i haven't been able to run it due to the
above.

the idea is to maintain a photo_indices array which goes from
0..NUMBER_OF_LINES and which is initially shuffled. current_photo_index
now gets incremented by one after every correct choice.  and when
current_photo_index reaches NUMBER_OF_LINES, it gets reset to 0 and
photo_indices is re-shuffled.